Converted unit tests to int

+++ validation/main_test.c
@@ -6,7 +6,7 @@
 #include "../convolution.h"
 #include <string.h>
 
-int compare_matrix(double *m, double *expected, int h, int w, int channels) {
+int compare_matrix(int *m, int *expected, int h, int w, int channels) {
 	for (int k = 0; k < channels; ++k) {
 		for (int i = 0; i < h; ++i) {
 			for (int j = 0; j < w; ++j) {
@@ -20,26 +20,29 @@ int compare_matrix(double *m, double *expected, int h, int w, int channels) {
 	return 0;
 }
 
-void test_pad_image(double *img, double *expected, int h, int w) {
-	double *out = padd0_image(h, w, img); // need to free out
+void test_pad_image(int *img, int *expected, int h, int w) {
+	int *out = padd0_image(h, w, img); // need to free out
 	if (compare_matrix(out, expected, h, w, 3) == 1) {
 		printf("test_pad_image FAILED\n");
 	}
 	free(out);
 }
 
-void test_calc_RGB_energy(double *img, double *expected, int h, int w) {
-	double *out = malloc(h * w * sizeof(double));
-	double *padded = padd0_image(h, w, img);
+void test_calc_RGB_energy(int *img, int *expected, int h, int w) {
+	int *out = malloc(h * w * sizeof(int));
+	int *padded = padd0_image(h, w, img);
+	//print_matrix(padded, w+2, h+2, 3);
+
 	calc_RGB_energy(h+2, w+2, padded, out);
 	if (compare_matrix(out, expected, h, w, 1) == 1) {
 		printf("test_calc_RGB_energy FAILED\n");
 	}
+	//print_matrix(out, w, h, 1);
 	free(padded);
 	free(out);
 }
 
-void test_min_seam(double *img, double expected, int h, int w, int is_vertical) {
+void test_min_seam(int *img, int expected, int h, int w, int is_vertical) {
 	int *backtrack;
 	if (is_vertical) {
 		backtrack = (int *)malloc(h * sizeof(int));
@@ -47,15 +50,15 @@ void test_min_seam(double *img, double expected, int h, int w, int is_vertical) 
 		backtrack = (int *)malloc(w * sizeof(int));
 	}
 	
-	double min_cost = min_seam(h, w, img, is_vertical, backtrack);
+	int min_cost = min_seam(h, w, img, is_vertical, backtrack);
 	if (min_cost != expected) {
 		printf("test_min_seam FAILED\n");
 	}
 	free(backtrack);
 }
 
-void test_optimal_image(double *img, double *expected, int h, int w, int h_diff, int w_diff) {
-	double *out;
+void test_optimal_image(int *img, int *expected, int h, int w, int h_diff, int w_diff) {
+	int *out;
 	out = optimal_image(w, h, w_diff, h_diff, img);
 	if (compare_matrix(out, expected, h-h_diff, w-w_diff, 3) == 1) {
 		printf("test_optimal_image FAILED\n");
@@ -67,7 +70,7 @@ int main(int argc, char const *argv[]) {
 	
 	// ------------ TEST IMAGE 1 -------------
 	{
-	double *img;
+	int *img;
 	int width, height;
 	if (!load_image("unit_tests/input_small/test_3.png", &width, &height, &img)) {
 		printf("Cannot load image");
@@ -75,7 +78,7 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{
-		double expected[] = {
+		int expected[] = {
 							0,0,0,0,0,
 							0,180,184,179,0,
 							0,179,212,255,0,
@@ -98,7 +101,7 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{
-		double expected[] = {
+		int expected[] = {
 							3886, 2836,4102,
 							2858,362,2738,
 							3808,2998,4024
@@ -107,17 +110,17 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{	// test vertical seam
-		double expected = 6196;
+		int expected = 6196;
 		test_min_seam(img, expected, height, width, 1);
 	}
 
 	{ // test horizontal seam
-		double expected = 5958;
+		int expected = 5958;
 		test_min_seam(img, expected, height, width, 0);
 	}
 
 	{ // test 2nd horizontal seam
-		double img[] = {
+		int img[] = {
 							180,184,179,
 							162,170,220,
 
@@ -127,12 +130,12 @@ int main(int argc, char const *argv[]) {
 							241,241,239,
 							236,240,245
 											};
-		double expected = 9954;
+		int expected = 9954;
 		test_min_seam(img, expected, height - 1, width, 0);
 	}
 
 	{ // test 2nd vertical seam
-		double img[] = {
+		int img[] = {
 							180,179,
 							179,255,
 							162,220,
@@ -145,12 +148,12 @@ int main(int argc, char const *argv[]) {
 							238,240,
 							236,245
 											};
-		double expected = 10064;
+		int expected = 10064;
 		test_min_seam(img, expected, height, width - 1, 1);
 	}
 
 	{	// test edge case : remove 0 seam
-		double expected[] = {
+		int expected[] = {
 							180,184,179,
 							179,212,255,
 							162,170,220,
@@ -172,7 +175,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							180,179,
 							179,255,
 							162,220,
@@ -194,7 +197,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							180,184,179,
 							162,170,220,
 
@@ -213,7 +216,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							180,179,
 							162,220,
 
@@ -232,7 +235,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							162,170,179,
 
 							162,204,176,
@@ -248,7 +251,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							162,
 
 							162,
@@ -261,7 +264,7 @@ int main(int argc, char const *argv[]) {
 
 	// ------------ TEST IMAGE 2 -------------
 	{
-	double *img;
+	int *img;
 	int width, height;
 	if (!load_image("unit_tests/input_small/test_2.png", &width, &height, &img)) {
 		printf("Cannot load image");
@@ -269,7 +272,7 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{
-		double expected[] = {
+		int expected[] = {
 							0,0,0,0,0,
 							0,191,229,251,0,
 							0,162,203,238,0,
@@ -289,7 +292,7 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{
-		double expected[] = {
+		int expected[] = {
 							2108, 1856, 2086,
 							2092, 1898, 2234
 											};
@@ -297,17 +300,17 @@ int main(int argc, char const *argv[]) {
 	}
 
 	{	// test vertical seam
-		double expected = 3754;
+		int expected = 3754;
 		test_min_seam(img, expected, height, width, 1);
 	}
 
 	{ // test horizontal seam
-		double expected = 6034;
+		int expected = 6034;
 		test_min_seam(img, expected, height, width, 0);
 	}
 
 	{	// test removing a vertical seam
-		double expected[] = {
+		int expected[] = {
 							191,251,
 							162,238,
 
@@ -326,7 +329,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							191,203,238,
 
 							53,52,48,
@@ -342,7 +345,7 @@ int main(int argc, char const *argv[]) {
 			printf("Cannot load image");
 			return 1;
 		}
-		double expected[] = {
+		int expected[] = {
 							191,238,
 
 							53,48,
